Correct interpolation for drag force values
