+++ src/main/java/com/waterloorocketry/airbrakeplugin/AirbrakePlugin.java
@@ -2,7 +2,7 @@ package com.waterloorocketry.airbrakeplugin;
 
 import com.waterloorocketry.airbrakeplugin.airbrake.Airbrakes;
 
-import com.waterloorocketry.airbrakeplugin.airbrake.LinearInterpAirbrakes;
+import com.waterloorocketry.airbrakeplugin.airbrake.SimulatedAirbrakes;
 import com.waterloorocketry.airbrakeplugin.controller.AlwaysOpenController;
 import com.waterloorocketry.airbrakeplugin.controller.Controller;
 import com.waterloorocketry.airbrakeplugin.controller.PIDController;
@@ -61,7 +61,7 @@ public class AirbrakePlugin extends AbstractSimulationExtension {
             controller = new PIDController(getTargetApogee(), getKp(), getKi(), getKd());
         }
 
-        Airbrakes airbrakes = new LinearInterpAirbrakes(new double[] { 0.0, 1.0 }, new double[] { 0.5, 1.5 });
+        Airbrakes airbrakes = new SimulatedAirbrakes();
         conditions.getSimulationListenerList().add(new AirbrakePluginSimulationListener(airbrakes, controller));
     }
